(split)bug: 46963
Change 263482 converted MailItem.TYPE* constants to enum.  Missed a spot in ShareInfo.

https://reviewboard.eng.vmware.com/r/197244/

http://bugzilla.zimbra.com/show_bug.cgi?id=46963

Copied from Perforce
 Change: 267529

+++ src/java/com/zimbra/cs/account/ShareInfo.java
@@ -518,7 +518,7 @@ public class ShareInfo {
                         md.put(MD_FOLDER_PATH, folder.getPath());
 
                         // for the same reason, we encode the default view for the folder in our metadata
-                        md.put(MD_FOLDER_DEFAULT_VIEW, folder.getDefaultView());
+                        md.put(MD_FOLDER_DEFAULT_VIEW, folder.getDefaultView().toByte());
 
                         mGrants.add(md);
                     }
